# [ML] inference performance optimizations and refactor