--- a/x-pack/plugin/core/src/main/java/org/elasticsearch/xpack/core/ml/inference/InferenceToXContentCompressor.java
+++ b/x-pack/plugin/core/src/main/java/org/elasticsearch/xpack/core/ml/inference/InferenceToXContentCompressor.java
@@ -10,6 +10,8 @@ import org.elasticsearch.common.CheckedFunction;
 import org.elasticsearch.common.bytes.BytesArray;
 import org.elasticsearch.common.bytes.BytesReference;
 import org.elasticsearch.common.io.stream.BytesStreamOutput;
+import org.elasticsearch.common.unit.ByteSizeUnit;
+import org.elasticsearch.common.unit.ByteSizeValue;
 import org.elasticsearch.common.xcontent.LoggingDeprecationHandler;
 import org.elasticsearch.common.xcontent.NamedXContentRegistry;
 import org.elasticsearch.common.xcontent.ToXContentObject;
@@ -37,7 +39,7 @@ public final class InferenceToXContentCompressor {
     // Either 10% of the configured JVM heap, or 1 GB, which ever is smaller
     private static final long MAX_INFLATED_BYTES = Math.min(
         (long)((0.10) * JvmInfo.jvmInfo().getMem().getHeapMax().getBytes()),
-        1_000_000_000); // 1 gb maximum
+        new ByteSizeValue(1, ByteSizeUnit.GB).getBytes());
 
     private InferenceToXContentCompressor() {}
 
@@ -46,9 +48,9 @@ public final class InferenceToXContentCompressor {
         return deflate(reference);
     }
 
-    static <T> T inflate(String compressedString,
-                         CheckedFunction<XContentParser, T, IOException> parserFunction,
-                         NamedXContentRegistry xContentRegistry) throws IOException {
+    public static <T> T inflate(String compressedString,
+                                CheckedFunction<XContentParser, T, IOException> parserFunction,
+                                NamedXContentRegistry xContentRegistry) throws IOException {
         try(XContentParser parser = JsonXContent.jsonXContent.createParser(xContentRegistry,
             LoggingDeprecationHandler.INSTANCE,
             inflate(compressedString, MAX_INFLATED_BYTES))) {