--- a/x-pack/plugin/core/src/main/java/org/elasticsearch/xpack/core/ml/inference/results/RawInferenceResults.java
+++ b/x-pack/plugin/core/src/main/java/org/elasticsearch/xpack/core/ml/inference/results/RawInferenceResults.java
@@ -10,7 +10,6 @@ import org.elasticsearch.ingest.IngestDocument;
 
 import java.io.IOException;
 import java.util.Arrays;
-import java.util.Map;
 import java.util.Objects;
 
 public class RawInferenceResults implements InferenceResults {
@@ -18,9 +17,9 @@ public class RawInferenceResults implements InferenceResults {
     public static final String NAME = "raw";
 
     private final double[] value;
-    private final Map<String, double[]> featureImportance;
+    private final double[][] featureImportance;
 
-    public RawInferenceResults(double[] value, Map<String, double[]> featureImportance) {
+    public RawInferenceResults(double[] value, double[][] featureImportance) {
         this.value = value;
         this.featureImportance = featureImportance;
     }
@@ -29,7 +28,7 @@ public class RawInferenceResults implements InferenceResults {
         return value;
     }
 
-    public Map<String, double[]> getFeatureImportance() {
+    public double[][] getFeatureImportance() {
         return featureImportance;
     }
 
@@ -44,7 +43,7 @@ public class RawInferenceResults implements InferenceResults {
         if (object == null || getClass() != object.getClass()) { return false; }
         RawInferenceResults that = (RawInferenceResults) object;
         return Arrays.equals(value, that.value)
-            && Objects.equals(featureImportance, that.featureImportance);
+            && Arrays.deepEquals(featureImportance, that.featureImportance);
     }
 
     @Override