--- a/x-pack/plugin/core/src/main/java/org/elasticsearch/xpack/core/ml/inference/trainedmodel/ensemble/Ensemble.java
+++ b/x-pack/plugin/core/src/main/java/org/elasticsearch/xpack/core/ml/inference/trainedmodel/ensemble/Ensemble.java
@@ -11,21 +11,12 @@ import org.apache.lucene.util.RamUsageEstimator;
 import org.elasticsearch.Version;
 import org.elasticsearch.common.Nullable;
 import org.elasticsearch.common.ParseField;
-import org.elasticsearch.common.collect.Tuple;
 import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.common.io.stream.StreamOutput;
 import org.elasticsearch.common.xcontent.ObjectParser;
 import org.elasticsearch.common.xcontent.XContentBuilder;
 import org.elasticsearch.common.xcontent.XContentParser;
-import org.elasticsearch.xpack.core.ml.inference.results.ClassificationInferenceResults;
-import org.elasticsearch.xpack.core.ml.inference.results.InferenceResults;
-import org.elasticsearch.xpack.core.ml.inference.results.RawInferenceResults;
-import org.elasticsearch.xpack.core.ml.inference.results.RegressionInferenceResults;
-import org.elasticsearch.xpack.core.ml.inference.trainedmodel.ClassificationConfig;
-import org.elasticsearch.xpack.core.ml.inference.trainedmodel.InferenceConfig;
-import org.elasticsearch.xpack.core.ml.inference.trainedmodel.InferenceHelpers;
 import org.elasticsearch.xpack.core.ml.inference.trainedmodel.LenientlyParsedTrainedModel;
-import org.elasticsearch.xpack.core.ml.inference.trainedmodel.NullInferenceConfig;
 import org.elasticsearch.xpack.core.ml.inference.trainedmodel.StrictlyParsedTrainedModel;
 import org.elasticsearch.xpack.core.ml.inference.trainedmodel.TargetType;
 import org.elasticsearch.xpack.core.ml.inference.trainedmodel.TrainedModel;
@@ -37,21 +28,14 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Objects;
 import java.util.OptionalDouble;
-import java.util.stream.Collectors;
 
-import static org.elasticsearch.xpack.core.ml.inference.trainedmodel.InferenceHelpers.classificationLabel;
-import static org.elasticsearch.xpack.core.ml.inference.trainedmodel.InferenceHelpers.decodeFeatureImportances;
-import static org.elasticsearch.xpack.core.ml.inference.trainedmodel.InferenceHelpers.transformFeatureImportance;
 
 public class Ensemble implements LenientlyParsedTrainedModel, StrictlyParsedTrainedModel {
 
     private static final long SHALLOW_SIZE = RamUsageEstimator.shallowSizeOfInstance(Ensemble.class);
-    // TODO should we have regression/classification sub-classes that accept the builder?
     public static final ParseField NAME = new ParseField("ensemble");
     public static final ParseField FEATURE_NAMES = new ParseField("feature_names");
     public static final ParseField TRAINED_MODELS = new ParseField("trained_models");
@@ -94,7 +78,7 @@ public class Ensemble implements LenientlyParsedTrainedModel, StrictlyParsedTrai
         return LENIENT_PARSER.apply(parser, null).build();
     }
 
-    private final List<String> featureNames;
+    private String[] featureNames;
     private final List<TrainedModel> models;
     private final OutputAggregator outputAggregator;
     private final TargetType targetType;
@@ -107,7 +91,7 @@ public class Ensemble implements LenientlyParsedTrainedModel, StrictlyParsedTrai
              TargetType targetType,
              @Nullable List<String> classificationLabels,
              @Nullable double[] classificationWeights) {
-        this.featureNames = Collections.unmodifiableList(ExceptionsHelper.requireNonNull(featureNames, FEATURE_NAMES));
+        this.featureNames = ExceptionsHelper.requireNonNull(featureNames, FEATURE_NAMES).toArray(String[]::new);
         this.models = Collections.unmodifiableList(ExceptionsHelper.requireNonNull(models, TRAINED_MODELS));
         this.outputAggregator = ExceptionsHelper.requireNonNull(outputAggregator, AGGREGATE_OUTPUT);
         this.targetType = ExceptionsHelper.requireNonNull(targetType, TARGET_TYPE);
@@ -118,7 +102,7 @@ public class Ensemble implements LenientlyParsedTrainedModel, StrictlyParsedTrai
     }
 
     public Ensemble(StreamInput in) throws IOException {
-        this.featureNames = Collections.unmodifiableList(in.readStringList());
+        this.featureNames = in.readStringArray();
         this.models = Collections.unmodifiableList(in.readNamedWriteableList(TrainedModel.class));
         this.outputAggregator = in.readNamedWriteable(OutputAggregator.class);
         this.targetType = TargetType.fromStream(in);
@@ -135,67 +119,8 @@ public class Ensemble implements LenientlyParsedTrainedModel, StrictlyParsedTrai
     }
 
     @Override
-    public InferenceResults infer(Map<String, Object> fields, InferenceConfig config, Map<String, String> featureDecoderMap) {
-        if (config.isTargetTypeSupported(targetType) == false) {
-            throw ExceptionsHelper.badRequestException(
-                "Cannot infer using configuration for [{}] when model target_type is [{}]", config.getName(), targetType.toString());
-        }
-        double[][] inferenceResults = new double[this.models.size()][];
-        List<Map<String, double[]>> featureInfluence = new ArrayList<>();
-        int i = 0;
-        NullInferenceConfig subModelInferenceConfig = new NullInferenceConfig(config.requestingImportance());
-        for (TrainedModel model : models) {
-            InferenceResults result = model.infer(fields, subModelInferenceConfig, Collections.emptyMap());
-            assert result instanceof RawInferenceResults;
-            RawInferenceResults inferenceResult = (RawInferenceResults) result;
-            inferenceResults[i++] = inferenceResult.getValue();
-            if (config.requestingImportance()) {
-                featureInfluence.add(inferenceResult.getFeatureImportance());
-            }
-        }
-        double[] processed = outputAggregator.processValues(inferenceResults);
-        return buildResults(processed,
-            decodeFeatureImportances(featureDecoderMap, mergeFeatureImportances(featureInfluence)),
-            config);
-    }
-
-    @Override
     public TargetType targetType() {
         return targetType;
-    }
-
-    private InferenceResults buildResults(double[] processedInferences,
-                                          Map<String, double[]> featureInfluence,
-                                          InferenceConfig config) {
-        // Indicates that the config is useless and the caller just wants the raw value
-        if (config instanceof NullInferenceConfig) {
-            return new RawInferenceResults(
-                new double[] {outputAggregator.aggregate(processedInferences)},
-                featureInfluence);
-        }
-        switch(targetType) {
-            case REGRESSION:
-                return new RegressionInferenceResults(outputAggregator.aggregate(processedInferences),
-                    config,
-                    transformFeatureImportance(featureInfluence, null));
-            case CLASSIFICATION:
-                ClassificationConfig classificationConfig = (ClassificationConfig) config;
-                assert classificationWeights == null || processedInferences.length == classificationWeights.length;
-                // Adjust the probabilities according to the thresholds
-                Tuple<Integer, List<ClassificationInferenceResults.TopClassEntry>> topClasses = InferenceHelpers.topClasses(
-                    processedInferences,
-                    classificationLabels,
-                    classificationWeights,
-                    classificationConfig.getNumTopClasses(),
-                    classificationConfig.getPredictionFieldType());
-                return new ClassificationInferenceResults((double)topClasses.v1(),
-                    classificationLabel(topClasses.v1(), classificationLabels),
-                    topClasses.v2(),
-                    transformFeatureImportance(featureInfluence, classificationLabels),
-                    config);
-            default:
-                throw new UnsupportedOperationException("unsupported target_type [" + targetType + "] for inference on ensemble model");
-        }
     }
 
     @Override
@@ -205,7 +130,7 @@ public class Ensemble implements LenientlyParsedTrainedModel, StrictlyParsedTrai
 
     @Override
     public void writeTo(StreamOutput out) throws IOException {
-        out.writeStringCollection(featureNames);
+        out.writeStringArray(featureNames);
         out.writeNamedWriteableList(models);
         out.writeNamedWriteable(outputAggregator);
         targetType.writeTo(out);
@@ -227,7 +152,7 @@ public class Ensemble implements LenientlyParsedTrainedModel, StrictlyParsedTrai
     @Override
     public XContentBuilder toXContent(XContentBuilder builder, Params params) throws IOException {
         builder.startObject();
-        if (featureNames.isEmpty() == false) {
+        if (featureNames.length > 0) {
             builder.field(FEATURE_NAMES.getPreferredName(), featureNames);
         }
         NamedXContentObjectHelper.writeNamedObjects(builder, params, true, TRAINED_MODELS.getPreferredName(), models);
@@ -252,7 +177,7 @@ public class Ensemble implements LenientlyParsedTrainedModel, StrictlyParsedTrai
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
         Ensemble that = (Ensemble) o;
-        return Objects.equals(featureNames, that.featureNames)
+        return Arrays.equals(featureNames, that.featureNames)
             && Objects.equals(models, that.models)
             && Objects.equals(targetType, that.targetType)
             && Objects.equals(classificationLabels, that.classificationLabels)
@@ -262,7 +187,7 @@ public class Ensemble implements LenientlyParsedTrainedModel, StrictlyParsedTrai
 
     @Override
     public int hashCode() {
-        return Objects.hash(featureNames,
+        return Objects.hash(Arrays.hashCode(featureNames),
             models,
             outputAggregator,
             targetType,
@@ -313,28 +238,12 @@ public class Ensemble implements LenientlyParsedTrainedModel, StrictlyParsedTrai
         return (long)Math.ceil(avg.getAsDouble()) + 2 * (models.size() - 1);
     }
 
-    @Override
-    public boolean supportsFeatureImportance() {
-        return models.stream().allMatch(TrainedModel::supportsFeatureImportance);
+    public List<TrainedModel> getModels() {
+        return models;
     }
 
-    Map<String, double[]> featureImportance(Map<String, Object> fields) {
-        return featureImportance(fields, Collections.emptyMap());
-    }
-
-    @Override
-    public Map<String, double[]> featureImportance(Map<String, Object> fields, Map<String, String> featureDecoder) {
-        Map<String, double[]> collapsed = mergeFeatureImportances(models.stream()
-            .map(trainedModel -> trainedModel.featureImportance(fields, Collections.emptyMap()))
-            .collect(Collectors.toList()));
-        return decodeFeatureImportances(featureDecoder, collapsed);
-    }
-
-    private static Map<String, double[]> mergeFeatureImportances(List<Map<String, double[]>> featureImportances) {
-        return featureImportances.stream()
-            .collect(HashMap::new,
-                (a, b) -> b.forEach((k, v) -> a.merge(k, v, InferenceHelpers::sumDoubleArrays)),
-                Map::putAll);
+    public TargetType getTargetType() {
+        return targetType;
     }
 
     public static Builder builder() {
@@ -344,7 +253,7 @@ public class Ensemble implements LenientlyParsedTrainedModel, StrictlyParsedTrai
     @Override
     public long ramBytesUsed() {
         long size = SHALLOW_SIZE;
-        size += RamUsageEstimator.sizeOfCollection(featureNames);
+        size += RamUsageEstimator.sizeOf(featureNames);
         size += RamUsageEstimator.sizeOfCollection(classificationLabels);
         size += RamUsageEstimator.sizeOfCollection(models);
         if (classificationWeights != null) {