This is a major refactor of the underlying inference logic.

The main refactor is now we are separating the model configuration and 
the inference interfaces. 

This has the following benefits:
 - we can store extra things with the model that are not 
   necessary for inference (i.e. treenode split information gain)
 - we can optimize inference separate from model serialization and storage.
 - The user is oblivious to the optimizations (other than seeing the benefits).

A major part of this commit is removing all inference related methods from the
trained model configurations (ensemble, tree, etc.) and moving them to a new class.

This new class satisfies a new interface that is ONLY for inference.

The optimizations applied currently are:
- feature maps are flattened once
- feature extraction only happens once at the highest level 
  (improves inference + feature importance through put)
- Only storing what we need for inference + feature importance on heap