@benwtrent 2020-06-04
Some numbers from JMH. These were gathered by running inference against a larger tree ensemble model. Each measurement is a single inference call against data with various fields missing and various expected results.

Before optimizations:
```
Benchmark                   Mode  Cnt        Score        Error  Units
infer_1105142  avgt   60  4676118.771 ± 408004.627  ns/op //~4ms
infer_1163199  avgt   60  6572394.355 ± 448022.365  ns/op //~6ms
infer_1274687  avgt   60  5221690.355 ± 298235.541  ns/op //~5ms
infer_1286491  avgt   60  6057884.209 ± 503322.898  ns/op //~6ms
infer_1329864  avgt   60  3704593.746 ± 209046.926  ns/op //~3ms
infer_1345914  avgt   60  3526360.753 ± 338985.848  ns/op //~3ms
infer_147975   avgt   60  2478510.531 ±  37657.708  ns/op //~2ms
infer_182200   avgt   60  2577532.404 ± 114494.789  ns/op //~2ms
infer_424703   avgt   60  6766206.021 ± 131101.240  ns/op //~6ms
infer_672407   avgt   60  8523977.602 ± 323823.581  ns/op //~8ms
```
After optimizations:
```
infer_1105142  avgt   60  105574.327 ± 4315.656  ns/op
infer_1163199  avgt   60  117342.726 ± 3857.156  ns/op
infer_1274687  avgt   60  113382.842 ± 2490.623  ns/op
infer_1286491  avgt   60  110763.581 ±  842.259  ns/op
infer_1329864  avgt   60  108473.284 ± 4547.068  ns/op
infer_1345914  avgt   60  104551.520 ±  336.243  ns/op
infer_147975   avgt   60   87700.209 ± 2342.228  ns/op
infer_182200   avgt   60  101667.732 ± 4628.543  ns/op
infer_424703   avgt   60  103756.438 ± 2358.109  ns/op
infer_672407   avgt   60  105891.101 ± 6329.288  ns/op
```